--- Gemfile.devel
@@ -1,1 +1,0 @@
-gem "isodoc", :git => "https://github.com/metanorma/isodoc.git", :branch => "features/figure-class"
